Finished service locator, got some tasks working

+++ lib/ancor/provider/openstack_instance_service.rb
@@ -3,19 +3,33 @@ module Ancor
     class OpenStackInstanceService < InstanceService
       interacts_with :os_nova
 
+      STATE_ACTIVE = 'ACTIVE'
+      STATE_ERROR = 'ERROR'
+
       # @param [Fog::Compute::OpenStack] connection
       # @param [Instance] instance
       # @return [undefined]
       def create(connection, instance)
         # TODO Lock the instance for this operation
 
+        os_instance = find_instance connection, instance
+        if os_instance
+          return if os_instance.state == STATE_ACTIVE
+          if os_instance.state == STATE_ERROR
+            raise 'Instance exists, but is in an error state'
+          end
+        end
+
+        nics = build_nic_specifications instance
+
         options = {
           name: instance.name,
-          flavor_ref: instance.provider_details["flavor_id"],
-          image_ref: instance.provider_details["image_id"],
-          nics: instance.networks,
-          # security_groups: [security_group],
-          user_data: instance.provider_details["user_data"] #@obj_store
+          flavor_ref: instance.provider_details['flavor_id'],
+          image_ref: instance.provider_details['image_id'],
+          nics: nics,
+          security_groups: [instance.provider_details['secgroup_id']],
+          # TODO Add support for user data from an object store
+          user_data: instance.provider_details['user_data']
         }
         os_instance = connection.servers.create options
       end
@@ -24,13 +38,75 @@ module Ancor
       # @param [Instance] instance
       # @return [undefined]
       def delete(connection, instance)
-        # TODO Lock the instance for this operation
+        os_instance = find_instance connection, instance
 
-        os_instance = connection.servers.find do |i|
-          i.name == instance.name
+        if os_instance
+          os_instance.destroy
+        end
+      end
+
+      def error?(connection, instance)
+        os_instance = find_instance connection, instance
+
+        unless os_instance
+          raise 'Instance not found'
         end
 
-        os_instance.destroy
+        os_instance.state == STATE_ERROR
+      end
+
+      def exists?(connection, instance)
+        !!find_instance(connection, instance)
+      end
+
+      def active?(connection, instance)
+        os_instance = find_instance connection, instance
+
+        unless os_instance
+          raise 'Instance not found'
+        end
+
+        pp os_instance.state
+
+        os_instance.state == STATE_ACTIVE
+      end
+
+      private
+
+      # Creates an array of specifications used to attach an instance to its associated networks
+      #
+      # @param [Instance] instance
+      # @return [Array]
+      def build_nic_specifications(instance)
+        instance.interfaces.map do |interface|
+          network = interface.network
+          ip_address = interface.ip_address
+
+          specification = {
+            net_id: network.provider_details['network_id'],
+            # this is not limited to IPv4 addresses, Fog is just weird
+            v4_fixed_ip: ip_address
+          }
+
+          specification.stringify_keys
+        end
+      end
+
+      # @param [Fog::Compute::OpenStack] connection
+      # @param [Instance] instance
+      # @return [Fog::Compute::OpenStack::Server]
+      def find_instance(connection, instance)
+        id = instance.provider_details['instance_id']
+        os_instance = connection.servers.get id
+
+        unless os_instance
+          # Find instance by name
+          os_instance = connection.servers.find { |i|
+            i.name == instance.name
+          }
+        end
+
+        os_instance
       end
     end # OpenStackInstanceService
   end # Provider
